requirements.txt
SQLAlchemy
alembic
flask_cors
flake8
python-dotenv
